further bugfixing and polishing
