+++ pyinstaller/electron/error_logs.html
@@ -1,23 +1,20 @@
 <html>
-    <link rel="stylesheet" type="text/css" href="./styles.css">
-    <body style="overflow: auto; height: 100%;">
-	<div class="card" style="width:90%">
-		This window shows you the last 700 lines Logs of the specterApp. This also includes 
-        the Logs of specterd which are marked as such. It might give you hints on why specter is not coming up properly.
-        The best approach is to scroll to the bottom and then search upwards for errors.
-        You can find the logfile in [yourHomedirectory]/.specter/specterApp.log.
-		<pre><code id="specterapp-logs"></code></pre>
-		<br>
-	
-	</div>
+  <link rel="stylesheet" type="text/css" href="./output.css" />
+  <body style="overflow: auto; height: 100%">
+    <div class="card" style="width: 90%">
+      This window shows you the last 700 lines Logs of the specterApp. This also includes the Logs of specterd which are
+      marked as such. It might give you hints on why specter is not coming up properly. The best approach is to scroll to the
+      bottom and then search upwards for errors. You can find the logfile in [yourHomedirectory]/.specter/specterApp.log.
+      <pre><code id="specterapp-logs"></code></pre>
+      <br />
+    </div>
     <script>
-        const fs = require('fs')
-        const path = require('path')
-        const helpers = require('./helpers')
-        helpers.getSpecterAppLogs( (lines) => {
-            document.getElementById('specterapp-logs').innerText = lines
-        }) 
-
+      const fs = require('fs')
+      const path = require('path')
+      const helpers = require('./src/helpers')
+      helpers.getSpecterAppLogs((lines) => {
+        document.getElementById('specterapp-logs').innerText = lines
+      })
     </script>
-    </body>
+  </body>
 </html>
